+++ sportorg/gui/main.py
@@ -2,8 +2,6 @@ import logging
 import sys
 from multiprocessing import freeze_support
 
-from sportorg.modules.configs.configs import Config, ConfigFile
-
 try:
     from PySide6.QtWidgets import QApplication
 except ModuleNotFoundError:
@@ -44,7 +42,7 @@ class Application(metaclass=Singleton):
         freeze_support()
         log_rust_status()
         try:
-            self.load_settings()
+            settings.load_settings_on_startup()
         except Exception as e:
             logging.exception("Error loading settings: %s", str(e))
         theme.apply_theme(self.app, settings.SETTINGS.theme)
@@ -59,73 +57,6 @@ class Application(metaclass=Singleton):
         self.set_rent_cards()
         self.main_window.show_window()
         sys.exit(self.app.exec_())
-
-    @staticmethod
-    def load_settings():
-        _, exists = settings.load_settings_from_file()
-        Config().read()
-        if not exists:
-            settings.SETTINGS.app_check_updates = Config().configuration.get(
-                "check_updates", True
-            )
-            settings.SETTINGS.locale = Config().configuration.get(
-                "current_locale", "ru_RU"
-            )
-            settings.SETTINGS.logging_level = Config().configuration.get(
-                "logging_level", True
-            )
-            settings.SETTINGS.logging_window_row_count = Config().configuration.get(
-                "log_window_row_count", True
-            )
-            settings.SETTINGS.window_show_toolbar = Config().configuration.get(
-                "show_toolbar", True
-            )
-            settings.SETTINGS.window_geometry = Config().geometry.get("main", "01")
-            settings.SETTINGS.window_dialog_path = Config().parser.get(
-                ConfigFile.DIRECTORY, "dialog_default_dir", fallback=""
-            )
-            settings.SETTINGS.race_use_birthday = Config().configuration.get(
-                "use_birthday", False
-            )
-            settings.SETTINGS.templates_path = settings.sanitize_path(
-                "templates_path", Config().templates.get("directory", "")
-            )
-            settings.SETTINGS.file_autosave_interval = Config().configuration.get(
-                "autosave_interval", 0
-            )
-            settings.SETTINGS.file_save_in_utf8 = Config().configuration.get(
-                "save_in_utf8", True
-            )
-            settings.SETTINGS.file_save_in_gzip = Config().configuration.get(
-                "save_in_gzip", True
-            )
-            settings.SETTINGS.file_generate_srb = Config().configuration.get(
-                "generate_srb", True
-            )
-            settings.SETTINGS.file_open_recent_file = Config().configuration.get(
-                "open_recent_file", True
-            )
-            settings.SETTINGS.printer_main = Config().printer.get("main", "")
-            settings.SETTINGS.printer_split = Config().printer.get("split", "")
-            settings.SETTINGS.sound_enabled = Config().sound.get("enabled", False)
-            settings.SETTINGS.sound_successful_path = Config().sound.get(
-                "successful", config.sound_dir("ok.wav")
-            )
-            settings.SETTINGS.sound_unsuccessful_path = Config().sound.get(
-                "unsuccessful", config.sound_dir("failure.wav")
-            )
-            settings.SETTINGS.sound_rented_card_enabled = Config().sound.get(
-                "enabled_rented_card", True
-            )
-            settings.SETTINGS.sound_rented_card_path = Config().sound.get(
-                "rented_card", config.sound_dir("rented_card.wav")
-            )
-            settings.SETTINGS.sound_enter_number_path = Config().sound.get(
-                "enter_number", config.sound_dir("enter_number.wav")
-            )
-            settings.SETTINGS.ranking = Config().ranking.get_all() or {}
-            settings.SETTINGS.ranking_ardf = Config().ranking_ardf.get_all() or {}
-            settings.save_settings_to_file()
 
     @staticmethod
     def set_status_comments():